Moved impl of WaitCondition into BundleRegistry.h
